move exception and wait function to utils
